rename the sse package to encoders
It's not obvious what sse means. And we will need to add other encoders
to allow ranges with chunked encoding
