+++ encoders/sse.go
@@ -1,4 +1,4 @@
-package sse
+package encoders
 
 import (
 	"bytes"
@@ -11,17 +11,17 @@ const (
 	data = "data: %s\n"
 )
 
-type encoder struct {
+type sseEncoder struct {
 	reader io.Reader // stores the original reader
 	offset int64     // offset for Seek purposes
 }
 
-// NewEncoder creates a new sse encoder
-func NewEncoder(r io.Reader) io.Reader {
-	return &encoder{reader: r}
+// NewSSEEncoder creates a new server-sent event encoder
+func NewSSEEncoder(r io.Reader) io.Reader {
+	return &sseEncoder{reader: r}
 }
 
-func (r *encoder) Seek(offset int64, whence int) (n int64, err error) {
+func (r *sseEncoder) Seek(offset int64, whence int) (n int64, err error) {
 	if seeker, ok := r.reader.(io.ReadSeeker); ok {
 		r.offset, err = seeker.Seek(offset, whence)
 	} else {
@@ -37,7 +37,7 @@ func (r *encoder) Seek(offset int64, whence int) (n int64, err error) {
 // FIXME: this version is simplified and assumes
 // that len(p) is always greater than the potential
 // length of data to be read.
-func (r *encoder) Read(p []byte) (n int, err error) {
+func (r *sseEncoder) Read(p []byte) (n int, err error) {
 	n, err = r.reader.Read(p)
 
 	if n > 0 {
